googlemap/requirements.txt
Django==1.9
requests==2.9.1
Pillow==3.1.1